+++ SE102/Engine/Scene.h
@@ -33,7 +33,7 @@ public:
 
 	virtual void Load() = 0;
 	virtual void Unload() = 0;
-	virtual void Update(DWORD dt) = 0;
+	virtual void Update(float dt) = 0;
 	virtual void Render() = 0; 
 
 	LPGAMEOBJECT const GetPlayer() const { return player; }
