+++ SE102/GameObjects/NPC/FirePiranha.h
@@ -47,6 +47,6 @@ public:
         bottom = position.y + 16; 
     }
 
-	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects = NULL) override;
+	void Update(float dt, vector<LPGAMEOBJECT>* coObjects = NULL) override;
 	void Render() override;
 };
